pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "apiary"
version = "0.5.1"
description = "A user-friendly tool for seamless interaction with the CoopHive protocol, enabling decentralized execution of computational workloads and management of payments and storage."
authors = [
    { name = "Levi Rybalov" },
    { name = "Matteo Manzi" },
    { name = "疒奀" },
    { name = "Aayushi Jain" }
]
requires-python = ">=3.12"
readme = "README.md"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "pip>=24.2",
    "click>=8.0,<8.1.4",
    "docker>=7.1.0,<7.2.0",
    "colorlog>=6.6.0,<6.9.0",
    "uvicorn>=0.30.0,<0.31.0",
    "fastapi>=0.112.0,<0.115.0",
    "lighthouseweb3>=0.1.0,<0.2.0",
    "readwrite>=0.6.0, <1.0",
    "python-dotenv>=1.0.0,<2.0",
    "ipykernel>=6.29.0,<7.0.0",
    "matplotlib>=3.9"
]
packages = [{include = "apiary"}]

[project.scripts]
"apiary" = "apiary.__main__:cli"

[project.optional-dependencies]
dev = [
    "pre-commit>=3.0.0,<3.3.0",
    "isort[colors]>=5.11.0,<5.12.0",
    "ruff>=0.6.0, < 1.0.0",
    "pydocstyle>=6.3.0,<6.4.0",
    "pytest>=8.3.0,<8.4.0",
    "pydeps>=1.11.0,<1.12.0"
]

[tool.maturin]
features = ["pyo3/extension-module"]
module-name = "apiary.apiars"
